# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: 3D Marketplace API
  version: "0.1.0"
  description: AR-powered 3D product marketplace API
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/auth/oauth/{provider}/callback:
    get:
      operationId: oauthCallback
      summary: OAuth callback handler
      tags: [auth]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
  /v1/auth/oauth/{provider}/token:
    post:
      operationId: oauthTokenExchange
      summary: Mobile OAuth token exchange (iOS sends provider id_token)
      tags: [auth]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleTokenRequest"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/token/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token using refresh token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke refresh token (logout)
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Logged out
  /v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current authenticated user
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/purchases:
    get:
      operationId: getMyPurchases
      summary: List current user's purchase history
      tags: [auth]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Purchase list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/summary:
    get:
      operationId: getUserSummary
      summary: Get user summary stats
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required:
        - error

    # --- Upload ---
...
    ProductStatus:
      type: string
      enum: [FOR_SALE, RESERVED, SOLD_OUT]
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        price_cents:
          type: integer
        seller_id:
          type: string
          format: uuid
        published_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        seller_name:
          type: string
          default: ""
        seller_avatar_url:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProductStatus"
        likes_count:
          type: integer
          default: 0
        views_count:
          type: integer
          default: 0
        chat_count:
          type: integer
          default: 0
        seller_location_name:
          type: string
          nullable: true
        is_liked:
          type: boolean
          nullable: true
      required:
        - id
        - asset_id
        - title
        - price_cents
        - seller_id
        - created_at
...
    PurchaseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        price_cents:
          type: integer
        purchased_at:
          type: string
          format: date-time
        product:
          $ref: "#/components/schemas/ProductResponse"
      required:
        - id
        - product_id
        - buyer_id
        - price_cents
        - purchased_at
    PurchaseListResponse:
      type: object
      properties:
        purchases:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - purchases
        - total
        - page
        - limit
...
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          nullable: true
        token_type:
          type: string
        expires_in:
          type: integer
          nullable: true
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - access_token
        - token_type
        - user
    GoogleTokenRequest:
      type: object
      properties:
        id_token:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
    TokenRefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token
    TokenRefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        provider:
          type: string
        avatar_url:
          type: string
          nullable: true
        trust_score:
          type: number
          format: float
          default: 36.5
        location_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - provider
        - created_at
    UserSummaryResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        product_count:
          type: integer
        unread_messages:
          type: integer
      required:
        - user
        - product_count
        - unread_messages